requirements.txt
numpy==1.23.5
pandas==1.5.3
mysql-connector-python==8.0.33